backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
langgraph==0.1.4
qdrant-client==1.9.1
pydantic-settings==2.3.4
python-multipart==0.0.9
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
redis==5.0.7
pytest==8.2.2
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
tiktoken==0.7.0
PyPDF2==3.0.1
